semaphoring-post-wait

+++ RowDaBoat-x64barebones-d4e1c147f975/Kernel/include/intQueue.h
@@ -9,7 +9,7 @@
 
 typedef struct IntQueueCDT * IntQueue;
 
-IntQueue newQueue(size_t size);
+IntQueue newQueue(int size);
 void enqueue(IntQueue q, int value);
 int dequeue(IntQueue q);
 int isEmpty(IntQueue q);
